Fixed initial create db bug (#115)
