+++ source/Frontend/scripts/addl-entries-script.js
@@ -2,17 +2,17 @@ import { getDailyEntries } from "../../Backend/api/entries_api.js";
 
 // Add additional entries bar web component
 
-const DATE = document.querySelector("log-type").readLog.header;
-const ADDLENTRYBAR = document.createElement("entry-bar");
-const ADDLENTRIES = document.querySelector(".additional")
+// const DATE = document.querySelector("log-type").readLog.header;
+// const ADDLENTRYBAR = document.createElement("entry-bar");
+// const ADDLENTRIES = document.querySelector(".additional")
 
-let entriesList = await getDailyEntries(DATE);
-let keys = entriesList[0];
-let fetchedEntries = entriesList[1];
+// let entriesList = await getDailyEntries(DATE);
+// let keys = entriesList[0];
+// let fetchedEntries = entriesList[1];
 
-ADDLENTRYBAR.type = "initial";
-ADDLENTRYBAR.entries = formatEntries(fetchedEntries, keys);
-ADDLENTRIES.appendChild(ADDLENTRYBAR);
+// ADDLENTRYBAR.type = "initial";
+// ADDLENTRYBAR.entries = formatEntries(fetchedEntries, keys);
+// ADDLENTRIES.appendChild(ADDLENTRYBAR);
