+++ source/Frontend/scripts/script.js
@@ -8,10 +8,11 @@ const SIDENAVROOT  = SIDENAV.shadowRoot;
 // const today = new Date();
 // router.setState("daily-log", false, today, "first-load");
 // create database
-router.setState("daily-log", false, new Date(), "on-load");
+
 //FIX LATER: Decide where to put all first time functions
 document.addEventListener("DOMContentLoaded", function() {
 	createDB();
+	router.setState("daily-log", false, new Date(), "on-load");
 });
