+++ source/Frontend/scripts/weekly-nav-script.js
@@ -3,11 +3,11 @@ import { router } from './router.js';
 import { getDailyPriority } from "../../Backend/api/bullet_api.js";
 
 
-let today = new Date();
-await createWeeklyNav(today);
-let WEEKLYNAV = document.querySelector("weekly-nav");
-WEEKLYNAV.shadowRoot.querySelector("[class='week-container']").style.opacity = "1";
-WEEKLYNAV.shadowRoot.querySelector("[class='weekly-nav-title']").style.opacity = "1";
+// let today = new Date();
+// await createWeeklyNav(today);
+// let WEEKLYNAV = document.querySelector("weekly-nav");
+// WEEKLYNAV.shadowRoot.querySelector("[class='week-container']").style.opacity = "1";
+// WEEKLYNAV.shadowRoot.querySelector("[class='weekly-nav-title']").style.opacity = "1";
 /**
  * createWeeklyNav
  * Takes in a date and creates a weekly nav component from that date and 
@@ -42,6 +42,7 @@ export async function createWeeklyNav(date) {
         }
 
     });
+    return true;
 } /* createWeeklyNav */
